requirements.txt
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
requests==2.25.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.1
wrapt==1.12.1
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.7
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.1.1
Pillow==7.2.0
psycopg2-binary==2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.2.0